Reorganize, move gradmultiply to parts
